issue: 109 Checkpointing for HomeObject
