+++ src/lib/homestore_backend/tests/CMakeLists.txt
@@ -5,6 +5,7 @@ include_directories (BEFORE .)
 list(APPEND TEST_SOURCES
     hs_shard_tests.cpp
     hs_blob_tests.cpp
+    homeobj_cp_tests.cpp
     )
 
 add_library(homestore_tests OBJECT)
